+++ js-miniapp-sdk/CHANGELOG.md
@@ -1,9 +1,13 @@
 ## CHANGELOG
 
-### v1.2.0 (In progress)
+### 1.x.x (In progress)
 
 - **Feature:** Added support for requesting the load and display of Interstitial & Rewarded Ads by the host application [See here](README.MD#4-Show-Ads)
-- **Feature:** Added `MiniApp.requestCustomPermissions` for requesting `USER_NAME`, `PROFILE_PHOTO`, and `CONTACT_LIST` permissions.
+
+### v1.2.0 (2020-10-02)
+
+- **Feature:** Added `MiniApp.requestCustomPermissions` for requesting `USER_NAME`, `PROFILE_PHOTO`, and `CONTACT_LIST` permissions. [See here](README.md#3-Request-Permissions)
+- **Feature:** Added `MiniApp.shareInfo` for sharing content with other Apps. [See here](README.md#4-Share-Info)
 
 ### v1.1.0 (2020-7-21)